apps/account/serializers.py
# 3 party imports
from rest_framework import serializers

# aydoor imports
from .models import User


class UserSerializer(serializers.HyperlinkedModelSerializer):
    email = serializers.EmailField()

    class Meta:
        model = User
        fields = ('id', 'email', 'password')
        extra_kwargs = {'password': {'write_only': True}}